@@ -1,7 +1,7 @@
 {
   "name": "webtorrent-desktop",
   "description": "WebTorrent, the streaming torrent client. For Mac, Windows, and Linux.",
-  "version": "0.18.0",
+  "version": "0.20.0",
   "author": {
     "name": "WebTorrent, LLC",
     "email": "feross@webtorrent.io",
@@ -14,36 +14,37 @@
     "airplayer": "^2.0.0",
     "application-config": "^1.0.0",
     "arch": "^2.0.0",
-    "auto-launch": "^4.0.1",
+    "auto-launch": "^5.0.5",
     "bitfield": "^1.0.2",
-    "capture-frame": "^1.0.0",
-    "chromecasts": "^1.8.0",
-    "cp-file": "^4.0.1",
-    "create-torrent": "^3.24.5",
+    "capture-frame": "^2.0.0",
+    "chokidar": "^2.0.4",
+    "chromecasts": "^1.9.1",
+    "cp-file": "^7.0.0",
+    "create-torrent": "^3.33.0",
     "debounce": "^1.0.0",
     "deep-equal": "^1.0.1",
     "dlnacasts": "^0.1.0",
-    "drag-drop": "^2.12.1",
+    "drag-drop": "^4.1.0",
     "es6-error": "^4.0.0",
     "fn-getter": "^1.0.0",
-    "iso-639-1": "^1.2.1",
-    "languagedetect": "^1.1.1",
+    "iso-639-1": "^2.0.5",
+    "languagedetect": "^1.2.0",
     "location-history": "^1.0.0",
-    "material-ui": "^0.17.0",
+    "material-ui": "^0.20.2",
     "mkdirp": "^0.5.1",
-    "musicmetadata": "^2.0.2",
+    "music-metadata": "^3.6.1",
     "natsort": "^1.0.6",
     "network-address": "^1.1.0",
-    "parse-torrent": "^5.7.3",
+    "parse-torrent": "^6.0.1",
     "prettier-bytes": "^1.0.1",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
-    "react-tap-event-plugin": "^2.0.1",
+    "prop-types": "^15.6.2",
+    "react": "^16.5.2",
+    "react-dom": "^16.5.2",
     "rimraf": "^2.5.2",
     "run-parallel": "^1.1.6",
-    "semver": "^5.1.0",
+    "semver": "^6.0.0",
     "simple-concat": "^1.0.0",
-    "simple-get": "^2.0.0",
+    "simple-get": "^3.0.3",
     "srt-to-vtt": "^1.1.1",
     "vlc-command": "^1.0.1",
     "webtorrent": "0.x",
@@ -51,25 +52,26 @@
     "zero-fill": "^2.2.3"
   },
   "devDependencies": {
-    "buble": "^0.15.2",
+    "babel-eslint": "^9.0.0",
+    "buble": "^0.19.6",
     "cross-zip": "^2.0.1",
-    "depcheck": "^0.6.4",
-    "electron": "1.6.0",
-    "electron-osx-sign": "0.4.3",
+    "depcheck": "^0.7.2",
+    "electron": "^1.8.8",
+    "electron-osx-sign": "^0.4.11",
     "electron-packager": "~8.5.1",
-    "electron-winstaller": "~2.5.2",
-    "gh-release": "^2.0.3",
+    "electron-winstaller": "^2.6.4",
+    "gh-release": "^3.4.0",
     "minimist": "^1.2.0",
     "nobin-debian-installer": "0.0.10",
-    "nodemon": "^1.10.2",
-    "open": "0.0.5",
-    "plist": "^2.0.1",
+    "nodemon": "^1.18.8",
+    "opn": "^6.0.0",
+    "plist": "^3.0.1",
     "pngjs": "^3.0.0",
     "run-series": "^1.1.4",
     "spectron": "^3.3.0",
     "standard": "*",
-    "tape": "^4.6.0",
-    "walk-sync": "^0.3.1"
+    "tape": "^4.9.1",
+    "walk-sync": "^1.1.3"
   },
   "engines": {
     "node": ">=4.0.0"
@@ -104,9 +106,12 @@
     "package": "node ./bin/package.js",
     "prepublish": "npm run build",
     "start": "npm run build && electron .",
-    "test": "standard && depcheck --ignores=buble,nodemon,gh-release --ignore-dirs=build,dist && node ./bin/extra-lint.js",
+    "test": "standard && depcheck --ignores=standard,babel-eslint --ignore-dirs=build,dist && node ./bin/extra-lint.js",
     "test-integration": "npm run build && node ./test",
     "update-authors": "./bin/update-authors.sh",
     "watch": "nodemon --exec \"npm run start\" --ext js,css --ignore build/ --ignore dist/"
+  },
+  "standard": {
+    "parser": "babel-eslint"
   }
 }
