@@ -1,5 +1,6 @@
 const path = require('path')
 const React = require('react')
+const PropTypes = require('prop-types')
 
 const colors = require('material-ui/styles/colors')
 const Checkbox = require('material-ui/Checkbox').default
@@ -7,7 +8,7 @@ const RaisedButton = require('material-ui/RaisedButton').default
 const Heading = require('../components/heading')
 const PathSelector = require('../components/path-selector')
 
-const {dispatch} = require('../lib/dispatcher')
+const { dispatch } = require('../lib/dispatcher')
 const config = require('../../config')
 
 class PreferencesPage extends React.Component {
@@ -37,7 +38,7 @@ class PreferencesPage extends React.Component {
           }}
           onChange={this.handleDownloadPathChange}
           title='Download location'
-          value={this.props.state.unsaved.prefs.downloadPath} />
+          value={this.props.state.saved.prefs.downloadPath} />
       </Preference>
     )
   }
@@ -51,7 +52,7 @@ class PreferencesPage extends React.Component {
       <Preference>
         <Checkbox
           className='control'
-          checked={!this.props.state.unsaved.prefs.openExternalPlayer}
+          checked={!this.props.state.saved.prefs.openExternalPlayer}
           label={'Play torrent media files using WebTorrent'}
           onCheck={this.handleOpenExternalPlayerChange} />
       </Preference>
@@ -67,7 +68,7 @@ class PreferencesPage extends React.Component {
       <Preference>
         <Checkbox
           className='control'
-          checked={this.props.state.unsaved.prefs.highestPlaybackPriority}
+          checked={this.props.state.saved.prefs.highestPlaybackPriority}
           label={'Highest Playback Priority'}
           onCheck={this.handleHighestPlaybackPriorityChange}
         />
@@ -81,10 +82,10 @@ class PreferencesPage extends React.Component {
   }
 
   externalPlayerPathSelector () {
-    const playerPath = this.props.state.unsaved.prefs.externalPlayerPath
+    const playerPath = this.props.state.saved.prefs.externalPlayerPath
     const playerName = this.props.state.getExternalPlayerName()
 
-    const description = this.props.state.unsaved.prefs.openExternalPlayer
+    const description = this.props.state.saved.prefs.openExternalPlayer
       ? `Torrent media files will always play in ${playerName}.`
       : `Torrent media files will play in ${playerName} if WebTorrent cannot play them.`
 
@@ -108,8 +109,61 @@ class PreferencesPage extends React.Component {
     dispatch('updatePreferences', 'externalPlayerPath', filePath)
   }
 
+  autoAddTorrentsCheckbox () {
+    return (
+      <Preference>
+        <Checkbox
+          className='control'
+          checked={this.props.state.saved.prefs.autoAddTorrents}
+          label={'Watch for new .torrent files and add them immediately'}
+          onCheck={(e, value) => { this.handleAutoAddTorrentsChange(e, value) }}
+        />
+      </Preference>
+    )
+  }
+
+  handleAutoAddTorrentsChange (e, isChecked) {
+    const torrentsFolderPath = this.props.state.saved.prefs.torrentsFolderPath
+    if (isChecked && !torrentsFolderPath) {
+      alert('Select a torrents folder first.') // eslint-disable-line
+      e.preventDefault()
+      return
+    }
+
+    dispatch('updatePreferences', 'autoAddTorrents', isChecked)
+
+    if (isChecked) {
+      dispatch('startFolderWatcher', null)
+      return
+    }
+
+    dispatch('stopFolderWatcher', null)
+  }
+
+  torrentsFolderPathSelector () {
+    const torrentsFolderPath = this.props.state.saved.prefs.torrentsFolderPath
+
+    return (
+      <Preference>
+        <PathSelector
+          dialog={{
+            title: 'Select folder to watch for new torrents',
+            properties: [ 'openDirectory' ]
+          }}
+          displayValue={torrentsFolderPath || ''}
+          onChange={this.handletorrentsFolderPathChange}
+          title='Folder to watch'
+          value={torrentsFolderPath ? path.dirname(torrentsFolderPath) : null} />
+      </Preference>
+    )
+  }
+
+  handletorrentsFolderPathChange (filePath) {
+    dispatch('updatePreferences', 'torrentsFolderPath', filePath)
+  }
+
   setDefaultAppButton () {
-    const isFileHandler = this.props.state.unsaved.prefs.isFileHandler
+    const isFileHandler = this.props.state.saved.prefs.isFileHandler
     if (isFileHandler) {
       return (
         <Preference>
@@ -142,7 +196,7 @@ class PreferencesPage extends React.Component {
         <Preference>
           <Checkbox
             className='control'
-            checked={this.props.state.unsaved.prefs.startup}
+            checked={this.props.state.saved.prefs.startup}
             label={'Open WebTorrent on startup.'}
             onCheck={this.handleStartupChange}
           />
@@ -163,8 +217,10 @@ class PreferencesPage extends React.Component {
     }
     return (
       <div style={style}>
-        <PreferencesSection title='Downloads'>
+        <PreferencesSection title='Folders'>
           {this.downloadPathSelector()}
+          {this.autoAddTorrentsCheckbox()}
+          {this.torrentsFolderPathSelector()}
         </PreferencesSection>
         <PreferencesSection title='Playback'>
           {this.openExternalPlayerCheckbox()}
@@ -183,7 +239,7 @@ class PreferencesPage extends React.Component {
 class PreferencesSection extends React.Component {
   static get propTypes () {
     return {
-      title: React.PropTypes.string
+      title: PropTypes.string
     }
   }
 
