@@ -7,7 +7,7 @@ const path = require('path')
 let hasErrors = false
 
 // Find all Javascript source files
-const files = walkSync('src', {globs: ['**/*.js']})
+const files = walkSync('src', { globs: ['**/*.js'] })
 console.log('Running extra-lint on ' + files.length + ' files...')
 
 // Read each file, line by line
@@ -35,7 +35,7 @@ files.forEach(function (file) {
     }
 
     if (error) {
-      let name = path.basename(file)
+      const name = path.basename(file)
       console.log('%s:%d - %s:\n%s', name, i + 1, error, line)
       hasErrors = true
     }
