@@ -1,7 +1,7 @@
 const React = require('react')
 
 const ModalOKCancel = require('./modal-ok-cancel')
-const {dispatch, dispatcher} = require('../lib/dispatcher')
+const { dispatch, dispatcher } = require('../lib/dispatcher')
 
 module.exports = class RemoveTorrentModal extends React.Component {
   render () {
