@@ -13,8 +13,10 @@ module.exports = {
   setRate
 }
 
+const http = require('http')
+
 const config = require('../../config')
-const {CastingError} = require('./errors')
+const { CastingError } = require('./errors')
 
 // Lazy load these for a ~300ms improvement in startup time
 let airplayer, chromecasts, dlnacasts
@@ -81,7 +83,7 @@ function testPlayer (type) {
   }
 
   function getDevices () {
-    return [{name: type + '-1'}, {name: type + '-2'}]
+    return [{ name: type + '-1' }, { name: type + '-2' }]
   }
 
   function open () {}
@@ -130,22 +132,47 @@ function chromecastPlayer () {
     })
   }
 
+  function serveSubtitles (callback) {
+    const subtitles = state.playing.subtitles
+    const selectedSubtitle = subtitles.tracks[subtitles.selectedIndex]
+    if (!selectedSubtitle) {
+      callback()
+    } else {
+      ret.subServer = http.createServer((req, res) => {
+        res.writeHead(200, {
+          'Content-Type': 'text/vtt;charset=utf-8',
+          'Access-Control-Allow-Origin': '*',
+          'Transfer-Encoding': 'chunked'
+        })
+        res.end(Buffer.from(selectedSubtitle.buffer.substr(21), 'base64'))
+      }).listen(0, function () {
+        const port = ret.subServer.address().port
+        const subtitlesUrl = 'http://' + state.server.networkAddress + ':' + port + '/'
+        callback(subtitlesUrl)
+      })
+    }
+  }
+
   function open () {
     const torrentSummary = state.saved.torrents.find((x) => x.infoHash === state.playing.infoHash)
-    ret.device.play(state.server.networkURL + '/' + state.playing.fileIndex, {
-      type: 'video/mp4',
-      title: config.APP_NAME + ' - ' + torrentSummary.name
-    }, function (err) {
-      if (err) {
-        state.playing.location = 'local'
-        state.errors.push({
-          time: new Date().getTime(),
-          message: 'Could not connect to Chromecast. ' + err.message
-        })
-      } else {
-        state.playing.location = 'chromecast'
-      }
-      update()
+    serveSubtitles(function (subtitlesUrl) {
+      ret.device.play(state.server.networkURL + '/' + state.playing.fileIndex, {
+        type: 'video/mp4',
+        title: config.APP_NAME + ' - ' + torrentSummary.name,
+        subtitles: subtitlesUrl ? [subtitlesUrl] : [],
+        autoSubtitles: !!subtitlesUrl
+      }, function (err) {
+        if (err) {
+          state.playing.location = 'local'
+          state.errors.push({
+            time: new Date().getTime(),
+            message: 'Could not connect to Chromecast. ' + err.message
+          })
+        } else {
+          state.playing.location = 'chromecast'
+        }
+        update()
+      })
     })
   }
 
@@ -159,6 +186,9 @@ function chromecastPlayer () {
 
   function stop (callback) {
     ret.device.stop(callback)
+    if (ret.subServer) {
+      ret.subServer.close()
+    }
   }
 
   function status () {
@@ -399,11 +429,11 @@ function toggleMenu (location) {
   }
 
   // Show a menu
-  state.devices.castMenu = {location, devices}
+  state.devices.castMenu = { location, devices }
 }
 
 function selectDevice (index) {
-  const {location, devices} = state.devices.castMenu
+  const { location, devices } = state.devices.castMenu
 
   // Start casting
   const player = getPlayer(location)
