@@ -3,7 +3,7 @@ module.exports = {
 }
 
 const config = require('../../config')
-const {InvalidSoundNameError} = require('./errors')
+const { InvalidSoundNameError } = require('./errors')
 const path = require('path')
 
 const VOLUME = 0.25
