@@ -1,6 +1,6 @@
 const appConfig = require('application-config')('WebTorrent')
 const path = require('path')
-const {EventEmitter} = require('events')
+const { EventEmitter } = require('events')
 
 const config = require('../../config')
 
@@ -113,7 +113,6 @@ function setupStateSaved (cb) {
   const cpFile = require('cp-file')
   const fs = require('fs')
   const parseTorrent = require('parse-torrent')
-  const parallel = require('run-parallel')
 
   const saved = {
     prefs: {
@@ -121,7 +120,10 @@ function setupStateSaved (cb) {
       isFileHandler: false,
       openExternalPlayer: false,
       externalPlayerPath: null,
-      startup: false
+      startup: false,
+      autoAddTorrents: false,
+      torrentsFolderPath: '',
+      highestPlaybackPriority: true
     },
     torrents: config.DEFAULT_TORRENTS.map(createTorrentObject),
     torrentsToResume: [],
@@ -130,26 +132,25 @@ function setupStateSaved (cb) {
 
   const tasks = []
 
-  config.DEFAULT_TORRENTS.map(function (t, i) {
+  config.DEFAULT_TORRENTS.forEach((t, i) => {
     const infoHash = saved.torrents[i].infoHash
-    tasks.push(function (cb) {
+    tasks.push(
       cpFile(
         path.join(config.STATIC_PATH, t.posterFileName),
         path.join(config.POSTER_PATH, infoHash + path.extname(t.posterFileName))
-      ).then(cb).catch(cb)
-    })
-    tasks.push(function (cb) {
+      )
+    )
+    tasks.push(
       cpFile(
         path.join(config.STATIC_PATH, t.torrentFileName),
         path.join(config.TORRENT_PATH, infoHash + '.torrent')
-      ).then(cb).catch(cb)
-    })
+      )
+    )
   })
 
-  parallel(tasks, function (err) {
-    if (err) return cb(err)
-    cb(null, saved)
-  })
+  Promise.all(tasks)
+    .then(() => cb(null, saved))
+    .catch(err => cb(err))
 
   function createTorrentObject (t) {
     // TODO: Doing several fs.readFileSync calls during first startup is not ideal
@@ -235,7 +236,7 @@ function saveImmediate (state, cb) {
     .filter((x) => x.infoHash)
     .map(function (x) {
       const torrent = {}
-      for (let key in x) {
+      for (const key in x) {
         if (key === 'progress' || key === 'torrentKey') {
           continue // Don't save progress info or key for the webtorrent process
         }
