@@ -22,7 +22,7 @@ test('torrent-list: show download path missing', function (t) {
     .then((windowTitle) => t.equal(windowTitle, 'Preferences', 'window title'))
     .then(() => setup.screenshotCreateOrCompare(app, t, 'prefs-basic'))
     .then(() => setup.endTest(app, t),
-          (err) => setup.endTest(app, t, err || 'error'))
+      (err) => setup.endTest(app, t, err || 'error'))
 })
 
 test('torrent-list: start, stop, and delete torrents', function (t) {
@@ -31,22 +31,20 @@ test('torrent-list: start, stop, and delete torrents', function (t) {
   const app = setup.createApp()
   setup.waitForLoad(app, t)
     .then(() => app.client.waitUntilTextExists('.torrent-list', 'Big Buck Bunny'))
-    // Mouse over the first torrent
-    .then(() => app.client.moveToObject('.torrent'))
-    .then(() => setup.screenshotCreateOrCompare(app, t, 'torrent-list-hover'))
+    .then(() => setup.screenshotCreateOrCompare(app, t, 'torrent-list'))
     // Click download on the first torrent, start downloading
     .then(() => app.client.click('.download input'))
     .then(() => app.client.waitUntilTextExists('.torrent-list', '276 MB'))
     .then(() => setup.screenshotCreateOrCompare(app, t, 'torrent-list-start-download'))
     // Click download on the first torrent again, stop downloading
     .then(() => app.client.click('.download input'))
-    .then(() => setup.screenshotCreateOrCompare(app, t, 'torrent-list-hover-download'))
+    .then(() => setup.screenshotCreateOrCompare(app, t, 'torrent-list-download'))
     // Click delete on the first torrent
     .then(() => app.client.click('.icon.delete'))
     .then(() => setup.screenshotCreateOrCompare(app, t, 'torrent-list-delete-prompt'))
     // Click cancel on the resulting confirmation dialog. Should be same as before.
     .then(() => app.client.click('.control.cancel'))
-    .then(() => setup.screenshotCreateOrCompare(app, t, 'torrent-list-hover'))
+    .then(() => setup.screenshotCreateOrCompare(app, t, 'torrent-list'))
     // Click delete on the first torrent again
     .then(() => app.client.click('.icon.delete'))
     .then(() => setup.screenshotCreateOrCompare(app, t, 'torrent-list-delete-prompt'))
@@ -54,7 +52,7 @@ test('torrent-list: start, stop, and delete torrents', function (t) {
     .then(() => app.client.click('.control.ok'))
     .then(() => setup.screenshotCreateOrCompare(app, t, 'torrent-list-deleted'))
     .then(() => setup.endTest(app, t),
-          (err) => setup.endTest(app, t, err || 'error'))
+      (err) => setup.endTest(app, t, err || 'error'))
 })
 
 test('torrent-list: expand torrent, unselect file', function (t) {
@@ -63,9 +61,6 @@ test('torrent-list: expand torrent, unselect file', function (t) {
   const app = setup.createApp()
   setup.waitForLoad(app, t)
     .then(() => app.client.waitUntilTextExists('.torrent-list', 'Big Buck Bunny'))
-    // Mouse over the torrent
-    .then(() => app.client.moveToObject('#torrent-cosmos'))
-    .then(() => setup.screenshotCreateOrCompare(app, t, 'torrent-list-cosmos-hover'))
     // Click on the torrent, expand
     .then(() => app.client.click('#torrent-cosmos'))
     .then(() => setup.screenshotCreateOrCompare(app, t, 'torrent-list-cosmos-expand'))
@@ -101,5 +96,5 @@ test('torrent-list: expand torrent, unselect file', function (t) {
     // Make sure that all the files are gone
     .then(() => setup.compareDownloadFolder(t, 'CosmosLaundromatFirstCycle', null))
     .then(() => setup.endTest(app, t),
-          (err) => setup.endTest(app, t, err || 'error'))
+      (err) => setup.endTest(app, t, err || 'error'))
 })
